feat: make point based on tuple of two floats

+++ src/common/models/stream_video.py
@@ -17,10 +17,10 @@ class VideoStreamRequestMessage:
   bottom_left_corner: Point
   top_right_corner: Point
 
-  def get_view_height(self) -> int:
+  def get_view_height(self) -> float:
     return abs(self.top_right_corner.y - self.bottom_left_corner.y)
 
-  def get_view_width(self) -> int:
+  def get_view_width(self) -> float:
     return abs(self.top_right_corner.x - self.bottom_left_corner.x)
